Move validation package to internal/mirror
- Package is only used by mirror pull/push, no need to keep it in pkg
