+++ internal/mirror/validation/registry_access.go
