+++ internal/mirror/validation/registry_access_test.go
